tracing: separate custom tag interface from trace driver
